requirements.txt
agno
openai
chromadb
pypdf
groq
python-dotenv
streamlit